MR-69: [WIP] Shared components implemented
