+++ src/Components/Layout/index.js
@@ -7,7 +7,7 @@ import SuperAdmins from '../SuperAdmins';
 import Trainers from '../Trainers';
 
 //Member
-import memberForm from '../Members/MemberForm';
+import MemberForm from '../Members/MemberForm';
 
 import Home from '../Home/index';
 import Header from '../Header/index';
@@ -27,9 +27,8 @@ function Layout() {
         <Route path="/admins" component={Admins} />
         <Route path="/classes" component={Classes} />
         <Route exact path="/members" component={Members} />
-        <Route exact path="/members/table" component={Members} />
-        <Route exact path="/members/form" component={memberForm} />
-        <Route path="/members/form/:id" component={memberForm} />
+        <Route exact path="/members/form" component={MemberForm} />
+        <Route path="/members/form/:id" component={MemberForm} />
         <Route path="/subscriptions" component={Subscriptions} />
         <Route path="/super-admins" component={SuperAdmins} />
         <Route path="/trainers" component={Trainers} />
